+++ packages/artrelay_flutter/lib/artrelay_flutter.dart
@@ -2,5 +2,7 @@
 library;
 
 export 'src/app_links_callback_source.dart';
+export 'src/artrelay_oauth_client.dart';
+export 'src/secure_pending_authorization_store.dart';
 export 'src/secure_token_store.dart';
 export 'src/system_uri_launcher.dart';